-ADD CompanyView Folder -ADD files to it
