+++ HQTCSDL.TeamProject.RealEstateAgency/View/CompanyView/AgenciesContent.xaml
@@ -1,11 +1,11 @@
 ﻿<UserControl
-    x:Class="HQTCSDL.TeamProject.RealEstateAgency.View.AgenciesContent"
+    x:Class="HQTCSDL.TeamProject.RealEstateAgency.View.CompanyView.AgenciesContent"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:controls="clr-namespace:MaterialDesignExtensions.Controls;assembly=MaterialDesignExtensions"
     xmlns:converter="clr-namespace:HQTCSDL.TeamProject.RealEstateAgency.Converters"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="clr-namespace:HQTCSDL.TeamProject.RealEstateAgency.View"
+    xmlns:local="clr-namespace:HQTCSDL.TeamProject.RealEstateAgency.View.CompanyView"
     xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:viewModel="clr-namespace:HQTCSDL.TeamProject.RealEstateAgency.ViewModel"
