+++ HQTCSDL.TeamProject.RealEstateAgency/View/CompanyView/AgenciesContent.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using MaterialDesignExtensions.Controls;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -13,14 +14,14 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace HQTCSDL.TeamProject.RealEstateAgency.View
+namespace HQTCSDL.TeamProject.RealEstateAgency.View.CompanyView
 {
     /// <summary>
-    /// Interaction logic for FlyoutContent.xaml
+    /// Interaction logic for AgenciesContent.xaml
     /// </summary>
-    public partial class FlyoutContent : UserControl
+    public partial class AgenciesContent : UserControl
     {
-        public FlyoutContent()
+        public AgenciesContent()
         {
             InitializeComponent();
         }
