+++ HQTCSDL.TeamProject.RealEstateAgency/View/CompanyView/FlyoutContent.xaml
@@ -1,9 +1,9 @@
 ﻿<UserControl
-    x:Class="HQTCSDL.TeamProject.RealEstateAgency.View.FlyoutContent"
+    x:Class="HQTCSDL.TeamProject.RealEstateAgency.View.CompanyView.FlyoutContent"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="clr-namespace:HQTCSDL.TeamProject.RealEstateAgency.View"
+    xmlns:local="clr-namespace:HQTCSDL.TeamProject.RealEstateAgency.View.CompanyView"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:wpf="http://materialdesigninxaml.net/winfx/xaml/themes"
     d:DesignHeight="300"
