+++ HQTCSDL.TeamProject.RealEstateAgency/View/CompanyView/HomeScreen.xaml
@@ -1,10 +1,10 @@
 ﻿<metro:MetroWindow
-    x:Class="HQTCSDL.TeamProject.RealEstateAgency.View.HomeScreen"
+    x:Class="HQTCSDL.TeamProject.RealEstateAgency.View.CompanyView.HomeScreen"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:iconPacks="http://metro.mahapps.com/winfx/xaml/iconpacks"
-    xmlns:local="clr-namespace:HQTCSDL.TeamProject.RealEstateAgency.View"
+    xmlns:local="clr-namespace:HQTCSDL.TeamProject.RealEstateAgency.View.CompanyView"
     xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:metro="http://metro.mahapps.com/winfx/xaml/controls"
